this PR reverts https://github.com/rack/rack/commit/e6b65296a247e23742d85e47cf31582b5883918a that was introduced for ruby 1.9 compatibility. The current version of rack does support ruby 2.2+

on ruby 2.4+ `unpack` is replaced by `unpack1` that doesn't allocate a temporarly array

       old decode      3.363M (± 4.9%) i/s -     16.775M in   5.005186s
       new decode      5.066M (± 5.8%) i/s -     25.224M in   5.004543s
       old encode      2.197M (± 5.4%) i/s -     10.970M in   5.016768s
       new encode      2.258M (± 4.5%) i/s -     11.277M in   5.007045s

benchmark
```
require 'base64'
require 'benchmark/ips'

str = 'test string'
unp = str.unpack("m*").first

Benchmark.ips do |x|
  x.report('old decode') { str.unpack("m*").first }
  x.report('new decode') { Base64.decode64(str) }
  x.report('old encode') { [unp].pack("m*") }
  x.report('new encode') { Base64.encode64(unp) }
  x.compare!
end
```